fix the bug for directional light. scene5 is working

+++ DirectionalLight.h
@@ -13,8 +13,8 @@ class DirectionalLight: public Light
 public:
     DirectionalLight();
     // ~DirectionalLight();
-    DirectionalLight(int x, int y, int z, float r, float g, float b);
-    DirectionalLight(int x, int y, int z, float r, float g, float b, float constant, float linear, float quadratic);
+    DirectionalLight(float x, float y, float z, float r, float g, float b);
+    DirectionalLight(float x, float y, float z, float r, float g, float b, float constant, float linear, float quadratic);
     Color getColor();
     bool getDir();
     Point getPos();
